Modified rust to import manga
