+++ package.json
@@ -38,6 +38,7 @@
     "jotai-optics": "^0.4.0",
     "lucide-react": "^0.487.0",
     "motion": "^12.6.3",
+    "nanoid": "^5.1.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router": "^7.5.0",
