+++ src/screens/Library.jsx
@@ -1,57 +1,52 @@
-import React, { useState, useEffect } from "react";
-import { Link, useNavigate } from "react-router";
-import { Origami, BookOpen, FolderPlus } from "lucide-react";
-import { Button } from "@/components/ui/button";
+import React, {useEffect, useState} from "react";
+import {useNavigate} from "react-router";
+import {Origami, Plus} from "lucide-react";
+import {nanoid} from "nanoid";
+import {open} from '@tauri-apps/plugin-dialog';
+import {readDir} from '@tauri-apps/plugin-fs';
+import {convertFileSrc, invoke} from '@tauri-apps/api/core';
+import {useAtom, useAtomValue} from "jotai";
+import {focusAtom} from "jotai-optics";
+
+import {settingsAtom, saveSettingsAtom} from "@/store/settings";
+import {libraryAtom, saveLibraryAtom} from "@/store/library.js";
+
+import {Button} from "@/components/ui/button";
 import DraggableMenuBar from "@/components/library/DraggableMenuBar";
-import MangaGrid from "@/components/library/MangaGrid.jsx";
-import { open } from '@tauri-apps/plugin-dialog';
-import { readDir } from '@tauri-apps/plugin-fs';
-import { convertFileSrc } from '@tauri-apps/api/core';
-import { useAtom } from "jotai";
-import { focusAtom } from "jotai-optics";
-import { settingsAtom, saveSettingsAtom } from "@/store/settings";
+import MangaCard from "@/components/library/MangaCard.jsx";
 
 // Settings Jotai Atoms for categories
 const categoriesAtom = focusAtom(settingsAtom, optic => optic.prop("categories"));
 const defaultCategoryAtom = focusAtom(settingsAtom, optic => optic.prop("default_category"));
+const mangaAtom = focusAtom(libraryAtom, optic => optic.prop("manga"));
 
 function Library() {
-  const [mangaList, setMangaList] = useState([]);
+  const [mangaList, setMangaList] = useAtom(mangaAtom);
   const [isLoading, setIsLoading] = useState(false);
 
   // Get categories from settings
   const [categories, setCategories] = useAtom(categoriesAtom);
   const [defaultCategory] = useAtom(defaultCategoryAtom);
   const [, saveSettings] = useAtom(saveSettingsAtom);
+  const [, saveLibrary] = useAtom(saveLibraryAtom);
+
+  useEffect( () => {
+    saveLibrary();
+  }, [mangaList]);
 
   // Use the default category from settings as the initial active category
-  const [activeCategory, setActiveCategory] = useState(defaultCategory);
   const [selectedCategory, setSelectedCategory] = useState(defaultCategory);
 
   const navigate = useNavigate();
 
-  // Load previously opened manga from storage
-  useEffect(() => {
-    const loadSavedManga = async () => {
-      try {
-        const savedManga = [];
-        setMangaList(savedManga);
-      } catch (error) {
-        console.error("Error loading saved manga:", error);
-      }
-    };
-
-    loadSavedManga();
-  }, []);
-
-  const addMangaFolder = async () => {
+  const handleAddManga = async () => {
     try {
       setIsLoading(true);
 
       const selected = await open({
         directory: true,
         multiple: false,
-        title: 'Select Manga Folder'
+        title: 'Select Source'
       });
 
       if (selected) {
@@ -59,7 +54,7 @@ function Library() {
         const folderName = folderPath.split('/').pop().split('\\').pop();
 
         // Try to find a cover image in the folder
-        const entries = await readDir(folderPath, { recursive: false });
+        const entries = await readDir(folderPath);
 
         // Filter for image files
         const imageExtensions = ['.jpg', '.jpeg', '.png', '.webp', '.gif', '.bmp'];
@@ -86,25 +81,21 @@ function Library() {
           });
 
         const coverImage = imageFiles.length > 0
-          ? convertFileSrc(folderPath+"/"+imageFiles[0].name)
+          ? convertFileSrc(folderPath + "/" + imageFiles[0].name)
           : 'https://placehold.co/200x300/png?text=No+Cover';
 
+
         // Create a new manga entry
-        const newManga = {
-          id: Date.now().toString(),
+        const mangaInput = {
           title: folderName,
           path: folderPath,
-          coverImage: coverImage,
-          lastRead: 'Just added',
-          progress: 0,
           category: selectedCategory,
-          totalPages: imageFiles.length
         };
 
-        // Add to our manga list
-        setMangaList(prev => [...prev, newManga]);
-      }
+        await invoke("import_manga_folder", {mangaInput})
 
+        // setMangaList([...mangaList, newManga]);
+      }
       setIsLoading(false);
     } catch (err) {
       console.error('Error adding manga:', err);
@@ -114,12 +105,11 @@ function Library() {
 
   const handleCategorySelect = (categoryId) => {
     setSelectedCategory(categoryId);
-    setActiveCategory(categoryId);
   };
 
   const handleMangaSelect = (manga) => {
     // Navigate to the reader with the manga path
-    navigate('/reader', { state: { mangaPath: manga.path } });
+    navigate('/reader', {state: {mangaPath: manga.path}});
   };
 
   // Handle categories reordering
@@ -129,7 +119,7 @@ function Library() {
   };
 
   // Filter manga by "all" or specific category
-  const filteredManga = selectedCategory === "all"
+  const filteredManga = selectedCategory === "all" && mangaList.length !== 0
     ? mangaList
     : mangaList.filter(manga => manga.category === selectedCategory);
 
@@ -146,37 +136,31 @@ function Library() {
       />
 
       {filteredManga.length > 0 ? (
-        <MangaGrid items={filteredManga} onMangaSelect={handleMangaSelect} />
+        <div className="flex flex-wrap gap-6 p-4 justify-center md:justify-start">
+          {filteredManga.map((manga) => (
+            <MangaCard
+              key={manga.id}
+              manga={manga}
+              onClick={handleMangaSelect}
+            />
+          ))}
+        </div>
       ) : (
         <div className="flex h-full w-full flex-1 flex-col justify-center m-auto gap-2 p-2 md:p-10">
           <div className="flex flex-col items-center">
-            <Origami size={200} strokeWidth={0.25} className="mb-6" />
+            <Origami size={200} strokeWidth={0.25} className="mb-6"/>
             <h1 className="mb-6 text-xl">Your library is empty</h1>
-
-            <div className="flex flex-col sm:flex-row gap-4">
-              <Button onClick={addMangaFolder} className="flex items-center gap-2">
-                <FolderPlus className="h-4 w-4" />
-                Add Manga Folder
-              </Button>
-
-              <Link to="/reader">
-                <Button variant="outline" className="flex items-center gap-2">
-                  <BookOpen className="h-4 w-4" />
-                  Open Reader
-                </Button>
-              </Link>
-            </div>
           </div>
         </div>
       )}
 
       {/* Floating action button to add manga */}
       <Button
-        onClick={addMangaFolder}
+        onClick={handleAddManga}
         className="fixed bottom-6 right-6 rounded-full h-14 w-14 shadow-lg"
         disabled={isLoading}
       >
-        <FolderPlus className="h-6 w-6" />
+        <Plus className="!h-6 !w-6"/>
       </Button>
     </div>
   );
